Change a logic of getting matching users
We don't check if rule broken anymore and  rely on checking individual conditions.
This is done to make sure that we can get a number of users for rules marked broken as
a result of error when executing SQL. E.g. extrnal DB is not avaliable.

Ideally we need to defer between these two ways of getting a rule broken.